Don't remove directory if exists, remove files in it instead
This resolves an edge case where the original WORKDIR is at `/workspace`, and then we proceed to download a directory from Nucleus to `/workspace`. This will cause the directory to be delete and recreated, making the current directory a deleted directory. If the user is not aware of this behavior, they may assume they are in the true `/workspace` directory.